int to np.int32

+++ tcxWeather.py
@@ -133,7 +133,7 @@ class TcxRide(object):
         print('Decimating to {0} Points'.format(numPoints))
         ind = np.linspace(0, (self.length-1), numPoints, endpoint=True, retstep=False, dtype=None)
         ind = np.floor(ind)
-        ind = ind.astype(int)
+        ind = ind.astype(np.int32)
 	print(ind)
         self.len = numPoints
         self.lat = np.array(self.latitude)
